pyproject.toml
[tool.poetry]
name = "user-form-crud"
version = "0.1.0"
description = ""
authors = ["Júlio Bernardes <juliocd.bernardes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.37.0"
psycopg = "^3.2.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
